feat: Refactor chat components and implement new ChatConversation structure
- Introduced ChatConversation component to encapsulate chat functionality.
- Created subcomponents: ConversationHeader, MessageListWindow, and ConversationInput for better modularity.
- Updated ChatSidebar to utilize new ChatList and SidebarHeader components.
- Enhanced MessageBubble for improved message display.
- Added NoWebSocketServer component for handling WebSocket connection issues.
- Removed deprecated ChatWindow component and adjusted routing accordingly.
- Refactored socketService to improve connection handling and added destroy method.
